chore: refactor JVM code
This moves the dicovery code for various API constructs into their own
classes, and makes the whole process more composable.
